fix: implement sandbox per platform to avoid landlock on android
