+++ sandbox.go
@@ -1,0 +1,4 @@
+package wireproxy
+
+type Sandbox struct {
+}
